requirements.txt
python-telegram-bot==20.5
openai==1.0.0